Shadow working for main icon. Pulling in the day of the week for Weekly forecast Still need to: Style collection view layout.

+++ Weather/Model/WeatherData.swift
@@ -35,7 +35,7 @@ struct Hourly: Codable {
 }
 
 struct Daily: Codable {
-    let dt: Int
+    let dt: Double
     let temp: TempData
     let weather: [Weather]
 }
